Move `TrimmedString` to utils.rs
