+++ core-backend/common/src/utils.rs
@@ -18,7 +18,10 @@
 
 use crate::StackEndError;
 use alloc::string::String;
+use codec::{Decode, Encode};
+use core::ops::Deref;
 use gear_core::memory::{PageU32Size, WasmPage};
+use scale_info::TypeInfo;
 
 #[macro_export]
 macro_rules! assert_ok {
@@ -41,7 +44,37 @@ macro_rules! assert_err {
     };
 }
 
-pub(crate) fn smart_truncate(s: &mut String, max_bytes: usize) {
+// Max amount of bytes allowed to be thrown as string explanation of the error.
+pub const TRIMMED_MAX_LEN: usize = 1024;
+
+/// Wrapped string to fit `core-backend::TRIMMED_MAX_LEN` amount of bytes.
+#[derive(
+    Decode, Encode, TypeInfo, Debug, Clone, PartialEq, Eq, PartialOrd, Ord, derive_more::Display,
+)]
+pub struct TrimmedString(String);
+
+impl TrimmedString {
+    pub(crate) fn new(mut string: String) -> Self {
+        smart_truncate(&mut string, TRIMMED_MAX_LEN);
+        Self(string)
+    }
+}
+
+impl<T: Into<String>> From<T> for TrimmedString {
+    fn from(other: T) -> Self {
+        Self::new(other.into())
+    }
+}
+
+impl Deref for TrimmedString {
+    type Target = String;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+fn smart_truncate(s: &mut String, max_bytes: usize) {
     let mut last_byte = max_bytes;
 
     if s.len() > last_byte {
